wallet tests

+++ src/visor/unconfirmed.go
@@ -43,7 +43,7 @@ func (self *UnconfirmedTxnPool) SetAnnounced(h coin.SHA256, t time.Time) {
 
 // Adds a coin.Transaction to the pool
 func (self *UnconfirmedTxnPool) RecordTxn(bc *coin.Blockchain,
-    t coin.Transaction, addrs map[coin.Address]int, didAnnounce bool) error {
+    t coin.Transaction, addrs map[coin.Address]byte, didAnnounce bool) error {
     if err := bc.VerifyTransaction(t); err != nil {
         return err
     }
